remove foreign keys constraint from CreateUsers migration

+++ db/migrate/20230529122744_create_users.rb
@@ -1,13 +1,7 @@
 class CreateUsers < ActiveRecord::Migration[7.0]
   def change
-    # Drop dependent objects first
-    remove_foreign_key :categories, column: :author_id
-    remove_foreign_key :purchases, column: :author_id
-    remove_foreign_key :category_purchases, :categories
-    remove_foreign_key :category_purchases, :purchases
-
     # Drop the users table
-    drop_table :users
+    drop_table :users if table_exists?(:users)
 
     # Recreate the users table with updated columns and constraints
     create_table :users do |t|
@@ -15,11 +9,5 @@ class CreateUsers < ActiveRecord::Migration[7.0]
 
       t.timestamps
     end
-
-    # Add back the foreign key constraints if needed
-    add_foreign_key :categories, :users, column: :author_id
-    add_foreign_key :purchases, :users, column: :author_id
-    add_foreign_key :category_purchases, :categories
-    add_foreign_key :category_purchases, :purchases
   end
 end
